Introduce micro_log.h to tensorflow/lite/micro
