+++ tensorflow/lite/micro/BUILD
@@ -250,6 +250,21 @@ cc_library(
 )
 
 cc_library(
+    name = "micro_log",
+    srcs = [
+        "micro_log.cc",
+    ],
+    hdrs = [
+        "micro_log.h",
+    ],
+    copts = micro_copts(),
+    deps = [
+        ":debug_log",
+        ":micro_string",
+    ],
+)
+
+cc_library(
     name = "micro_error_reporter",
     srcs = [
         "micro_error_reporter.cc",
@@ -259,9 +274,8 @@ cc_library(
     ],
     copts = micro_copts(),
     deps = [
-        ":debug_log",
         ":micro_compatibility",
-        ":micro_string",
+        ":micro_log",
         "//tensorflow/lite/core/api",
     ],
 )
