+++ tensorflow/lite/micro/micro_error_reporter.cc
@@ -19,38 +19,11 @@ limitations under the License.
 #include <cstdint>
 #include <new>
 
-#if !defined(TF_LITE_STRIP_ERROR_STRINGS)
-#include "tensorflow/lite/micro/debug_log.h"
-#include "tensorflow/lite/micro/micro_string.h"
-#endif
-
 namespace {
 uint8_t micro_error_reporter_buffer[sizeof(tflite::MicroErrorReporter)];
 tflite::MicroErrorReporter* error_reporter_ = nullptr;
 
-void Log(const char* format, va_list args) {
-#if !defined(TF_LITE_STRIP_ERROR_STRINGS)
-  // Only pulling in the implementation of this function for builds where we
-  // expect to make use of it to be extra cautious about not increasing the code
-  // size.
-  static constexpr int kMaxLogLen = 256;
-  char log_buffer[kMaxLogLen];
-  MicroVsnprintf(log_buffer, kMaxLogLen, format, args);
-  DebugLog(log_buffer);
-  DebugLog("\r\n");
-#endif
-}
-
 }  // namespace
-
-#if !defined(TF_LITE_STRIP_ERROR_STRINGS)
-void MicroPrintf(const char* format, ...) {
-  va_list args;
-  va_start(args, format);
-  Log(format, args);
-  va_end(args);
-}
-#endif
 
 namespace tflite {
 ErrorReporter* GetMicroErrorReporter() {
